% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{detections_on_path}
\alias{detections_on_path}
\title{Determine which transmissions get detected on a path}
\usage{
detections_on_path(
  path,
  transmissionTime,
  rec_geometry,
  rec_id,
  rec_deploy_time,
  rec_ret_time,
  detection_fn,
  fish_id,
  end_crs = NULL
)
}
\arguments{
\item{path}{the sf geometry from path_transmissions containing the points at each transmission time}

\item{transmissionTime}{the vector of transmission times from path_transmissions output}

\item{rec_geometry}{the sf geometry of points for receivers}

\item{rec_id}{the vector of receiver ids (e.g., Receiver)}

\item{rec_deploy_time}{the vector of receiver deployment times (e.g., DeployedDateTime)}

\item{rec_ret_time}{the vector of receiver retrieval times (e.g., RetrivedDateTime)}

\item{detection_fn}{a function giving the probability of a transmission being detected in metres}

\item{fish_id}{the id you want to add to the detections}

\item{end_crs}{(optional) the crs to use at the end so that you can transform the receiver and fish locations on output}
}
\description{
Determine which transmissions get detected on a path
}
