% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{get_time_surv}
\alias{get_time_surv}
\title{Calculate the survival needed to hit an annual survival target}
\usage{
get_time_surv(num_time_periods, annual_survival)
}
\arguments{
\item{num_time_periods}{the number of time periods (e.g, 365 for days)}

\item{annual_survival}{the annual survival}
}
\description{
Calculate the survival needed to hit an annual survival target
}
