% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{get_vitality_parm}
\alias{get_vitality_parm}
\title{Find Vitality 2009 parameters for battery survival}
\usage{
get_vitality_parm(
  days1,
  surv_per1,
  maxiter = 1000,
  numIslands = 8,
  upper_bound = c(5, 5, 5, 0.05)
)
}
\arguments{
\item{days1}{number of days for the first survival percent to be matched}

\item{surv_per1}{the first survival percent to be matched}

\item{maxiter}{the number of iterations for the genetic algoritim}

\item{numIslands}{the number of islands for the genetic algorithm to find the values}

\item{upper_bound}{the upper bound for the values of r, s, k and u}
}
\description{
Find Vitality 2009 parameters for battery survival
}
