% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{get_weibull_parm}
\alias{get_weibull_parm}
\title{Find Weibull distribution parameters for battery survival}
\usage{
get_weibull_parm(
  days,
  surv_per,
  init_parm = c(1, 1),
  lower_bound = c(0, 0),
  upper_bound = c(5, 5),
  maxiter = 1000,
  numIslands = 8
)
}
\arguments{
\item{days}{number of days for that survival percent to be matched}

\item{surv_per}{the survival percentage to hit}

\item{init_parm}{the initial parameters to use}
}
\description{
Assumes that shape must > 1
}
