% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{path_transmissions}
\alias{path_transmissions}
\title{Add transmissions to a path that has known timesteps at a location}
\usage{
path_transmissions(
  path,
  timesteps,
  delayRNG = c(60, 180),
  burstDur = 5,
  tz = "UTC"
)
}
\arguments{
\item{path}{the sf points to add transmissions}

\item{timesteps}{the POSIXct timesteps of the path}

\item{delayRNG}{the min/max delay time from transmission to the next}

\item{burstDur}{number of seconds for the duration of the transmission}

\item{tz}{timezone of the time steps}
}
\description{
Add transmissions to a path that has known timesteps at a location
}
