% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{plot_individual}
\alias{plot_individual}
\title{Plot an indivdual's path and attraction points}
\usage{
plot_individual(individual, ylim = NULL, xlim = NULL)
}
\arguments{
\item{individual}{output of simulate_individual}
}
\description{
Plot an indivdual's path and attraction points
}
