% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{pvitality}
\alias{pvitality}
\title{Vitality (2009 Li & Anderson) F dist}
\usage{
pvitality(q, r, s, k, u)
}
\arguments{
\item{q}{the quantile (time)}

\item{r}{wear rate}

\item{s}{standard deviation in wear rate}

\item{k}{rate of accidental failure}

\item{u}{standard deviation in accidental failure}
}
\description{
One minus the survival equation described
}
