% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{qvitality}
\alias{qvitality}
\title{Quantile of Vitality 2009}
\usage{
qvitality(p, r, s, k, u)
}
\arguments{
\item{p}{the target percentile}

\item{r}{wear rate}

\item{s}{standard deviation in wear rate}

\item{k}{rate of accidental failure}

\item{u}{standard deviation in accidental failure}
}
\description{
Quantile of Vitality 2009
}
