% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{rvitality}
\alias{rvitality}
\title{Random values of Vitality 2009}
\usage{
rvitality(n, r, s, k, u)
}
\arguments{
\item{n}{the number required}

\item{r}{wear rate}

\item{s}{standard deviation in wear rate}

\item{k}{rate of accidental failure}

\item{u}{standard deviation in accidental failure}
}
\description{
Random values of Vitality 2009
}
