% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{survival_function}
\alias{survival_function}
\title{Get the survival given some set of parameters}
\usage{
survival_function(F_dist, parameters, t_adj = 1000)
}
\arguments{
\item{F_dist}{the F distribution to get survival function of}

\item{parameters}{parameters of the distribution as a list}

\item{t_adj}{the amount to adjust t}
}
\description{
Get the survival given some set of parameters
}
