% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{weibull_h_gen}
\alias{weibull_h_gen}
\title{Generate the weibull hazard function for days with a given shape/scale}
\usage{
weibull_h_gen(shape, scale)
}
\arguments{
\item{shape}{the shape parameter}

\item{scale}{the scale parameter}
}
\description{
Generate the weibull hazard function for days with a given shape/scale
}
