% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{sim_ou_boundry}
\alias{sim_ou_boundry}
\title{Simulate a 2D Ornstein-Uhlenbeck process with boundary}
\usage{
sim_ou_boundry(
  polygon,
  timesteps,
  b,
  a = diag(0.5, 2),
  x0,
  sigma = matrix(c(50, 0, 50, 0), nrow = 2),
  speed = 2
)
}
\arguments{
\item{polygon}{the sf polygon boundry the movement should not cross (polygon crs should be in metres)}

\item{timesteps}{the timesteps of the movement}

\item{b}{vector or matrix of attraction point(s), if matrix each row is the attraction point at the timestep}

\item{a}{2x2 matrix controlling the force of the attraction}

\item{x0}{the intial start point}

\item{sigma}{the 2x2 matrix controlling the random noise in the movement}

\item{speed}{max speed in km/h}
}
\description{
This merges elements of glatos' crw_in_polygon and adeHabitatLT simm.mou. You can simulate
points corresponding to a Ornstein-Uhlenbeck process with an attraction point but that won't cross
polygon boundries. The boundries are avoided simply by trying to find another point that doesn't cause the path
to cross the boundry, so in some instances it can have trouble and will fail if it can't find a good point.
The attraction points can be set seperate for every time step potentially and so can move.
}
