% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsim.R
\name{simulate_individual}
\alias{simulate_individual}
\title{Simulate one individual's detections and track.}
\usage{
simulate_individual(
  fish_id,
  release_datetime,
  release_location,
  start_age,
  individual_s_fn,
  battery_s_fn,
  rec_info,
  raster,
  polygon,
  b_frame,
  ou_opts = NULL,
  start_crs,
  mid_crs,
  end_crs,
  end_time = as.POSIXct("2025-12-31 23:59:59", tz = "UTC"),
  delayRNG = c(60, 180),
  burstDur = 5,
  timestep_per_day = 0.75
)
}
\arguments{
\item{fish_id}{the ID to give the individual}

\item{release_datetime}{the release date time of the individual (POSIXct)}

\item{release_location}{vector of the release location of the individual}

\item{start_age}{the age at release (in days)}

\item{individual_s_fn}{the survival function of the individual (in days)}

\item{battery_s_fn}{the survival function of the transmitter battery (in days)}

\item{rec_info}{named list containing the rec_geometry, rec_id, rec_deploy_time, rec_ret_time, and detection_fn for detections_on_path}

\item{polygon}{the polygon for sim_ou_boundary (must have mid_crs)}

\item{b_frame}{the attrator points in the start_crs, data frame with x,y, and change_datetime for when that attrator point applies}

\item{ou_opts}{optional named list giving the parameters of sim_ou_boundry other than b and timesteps}

\item{start_crs}{the crs of the release_location}

\item{mid_crs}{the crs used for sim_ou_boundry}

\item{end_crs}{the final output crs}

\item{end_time}{the end date time of the simulation}

\item{delayRNG}{the min/max delay time from transmission to the next}

\item{burstDur}{number of seconds for the duration of the transmission}

\item{timestep_per_day}{the number of tries for sim_ou_boundry}
}
\description{
Simulate one individual's detections and track.
}
